Add ZipSliceArchive::into_zip_archive
For types that implement AsRef<[u8]> but not ReaderAt, the
`ZipSliceArchive::into_reader` method has dubious benefits as many
methods on `ZipArchive` require a `ReaderAt` (like calling `entries`).

This is why `into_reader` has been deprecated in favor of a new method:
`into_zip_archive` which guarantees a `ReaderAt` implementation. It does
require wrapping the reader in a Cursor, which doesn't feel great, but
better than a custom type.